+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog][docs-changelog], and the version adher
 
 ## Unreleased
 ### Added
+- Cluster-set cluster status enum.
 - Complete error message to the LocalExecutor raised exception
 ### Fixed
 - Possible password leak in LocalExecutor traceback info.